Added generic OrbitRenderObject and orbit menu so that all objects that have to edit their orbit or visualise them can use them. (#90)
